MAINT: Renames file for compaction analysis, as it is a much more genreal translation analysis.
