+++ src/daria/__init__.py
@@ -23,4 +23,4 @@ from daria.corrections.shape.homography import *
 from daria.corrections.shape.translation import *
 from daria.corrections.shape.piecewiseperspective import *
 from daria.corrections.color.colorchecker import *
-from daria.analysis.compaction import *
+from daria.analysis.translationanalysis import *
